refactor: Simplify setup of ApplicationRoot
