+++ packages/ui/src/components/Layout/ApplicationRoot.tsx
@@ -1,6 +1,7 @@
 import React, { ElementRef } from 'react'
 import { ComponentProps } from '@stitches/react'
-import { styled } from '~/theme'
+import { styled, ToggleThemeButton } from '~/theme'
+import { menuWidth } from '../SideMenu'
 
 const StyledApplicationRoot = styled('div', {
   position: 'relative',
@@ -14,7 +15,32 @@ const StyledApplicationRoot = styled('div', {
   overflowY: 'hidden',
 })
 
-export type ApplicationRootProps = {}
+const TopRight = styled('div', {
+  position: 'absolute',
+  top: 0,
+  right: 0,
+  paddingTop: 10,
+  paddingRight: 16,
+})
+
+const PageWrapper = styled('div', {
+  position: 'relative',
+  width: '100%',
+  height: '100%',
+  backgroundColor: '$Background2dp',
+
+  variants: {
+    sideMenu: {
+      true: {
+        paddingLeft: menuWidth,
+      },
+    },
+  },
+})
+
+export type ApplicationRootProps = {
+  navigation?: React.ReactElement
+}
 
 type ForwardProps = ComponentProps<typeof StyledApplicationRoot> &
   ApplicationRootProps
@@ -23,11 +49,19 @@ export const ApplicationRoot = React.forwardRef<
   ElementRef<typeof StyledApplicationRoot>,
   ForwardProps
 >((properties, forwardedRef) => {
-  const { children, ...remainingProps } = properties
+  const { children, navigation, ...remainingProps } = properties
+
+  const hasSideMenu = Boolean(navigation)
 
   return (
     <StyledApplicationRoot ref={forwardedRef} {...remainingProps}>
-      {children}
+      {navigation}
+
+      <PageWrapper sideMenu={hasSideMenu}>{children}</PageWrapper>
+
+      <TopRight>
+        <ToggleThemeButton />
+      </TopRight>
     </StyledApplicationRoot>
   )
 })
