+++ packages/ui/src/components/Layout/Page.tsx
@@ -1,22 +1,6 @@
 import React, { ElementRef } from 'react'
 import { ComponentProps } from '@stitches/react'
 import { styled } from '~/theme'
-import { menuWidth } from '~/components/SideMenu'
-
-const PageWrapper = styled('div', {
-  position: 'relative',
-  width: '100%',
-  height: '100%',
-  backgroundColor: '$Background2dp',
-
-  variants: {
-    sideMenu: {
-      true: {
-        paddingLeft: menuWidth,
-      },
-    },
-  },
-})
 
 const StyledPage = styled('div', {
   display: 'flex',
@@ -31,15 +15,12 @@ const StyledPage = styled('div', {
         margin: '0 auto',
         maxWidth: '860px',
       },
-
-      fullscreen: {},
     },
   },
 })
 
 export type PageProps = {
-  mode?: 'center' | 'fullscreen'
-  hasSideMenu?: boolean
+  mode?: 'center'
 }
 
 type ForwardProps = ComponentProps<typeof StyledPage> & PageProps
@@ -48,13 +29,11 @@ export const Page = React.forwardRef<
   ElementRef<typeof StyledPage>,
   ForwardProps
 >((properties, forwardedRef) => {
-  const { mode, hasSideMenu = true, children, ...remainingProps } = properties
+  const { mode, children, ...remainingProps } = properties
 
   return (
-    <PageWrapper sideMenu={hasSideMenu}>
-      <StyledPage mode={mode} ref={forwardedRef} {...remainingProps}>
-        {children}
-      </StyledPage>
-    </PageWrapper>
+    <StyledPage mode={mode} ref={forwardedRef} {...remainingProps}>
+      {children}
+    </StyledPage>
   )
 })
